minor fix to fetch string value information from older archives also just like objects, since they could have been created much earlier and then reused later

+++ src/main/java/com/insidious/plugin/extension/model/ReplayData.java
@@ -299,7 +299,7 @@ public class ReplayData {
         return stringInfoMap.get(String.valueOf(id));
     }
 
-    public ObjectInfo getObjectInfo(int id) {
+    public ObjectInfo getObjectInfo(long id) {
         return objectInfoMap.get(String.valueOf(id));
     }
